Move config singleton/Remove unnessecary function
