+++ valhalla/config.h
@@ -1,6 +1,14 @@
-/* Minimal valhalla/config.h template for minimal CMake build configuration */
+#ifndef VALHALLA_CONFIG_H_
+#define VALHALLA_CONFIG_H_
 
-#include <valhalla/valhalla.h>
+#include <stdexcept>
+#include <string>
+#include <filesystem>
+
+#include <boost/property_tree/ptree.hpp>
+
+#include "baldr/rapidjson_utils.h"
+#include "valhalla.h"
 
 #define VALHALLA_STRINGIZE_NX(A) #A
 #define VALHALLA_STRINGIZE(A) VALHALLA_STRINGIZE_NX(A)
@@ -31,3 +39,42 @@
 
 /* Define to the version of this package. */
 #define PACKAGE_VERSION VALHALLA_VERSION
+
+namespace valhalla{
+const boost::property_tree::ptree& config(const std::string&);
+}
+
+namespace {
+struct config_singleton_t {
+protected:
+  boost::property_tree::ptree config_;
+
+  config_singleton_t() = delete;
+  config_singleton_t(const std::string& config_file_or_inline) {
+    if (config_file_or_inline.empty()) {
+      throw std::runtime_error("no config provided");
+    }
+
+    if (std::filesystem::is_regular_file(config_file_or_inline)) {
+      rapidjson::read_json(config_file_or_inline, config_);
+    } else {
+      auto inline_config = std::stringstream(config_file_or_inline);
+      rapidjson::read_json(inline_config, config_);
+    }
+  }
+
+public:
+  config_singleton_t(config_singleton_t const&) = delete;
+  void operator=(const config_singleton_t&) = delete;
+  friend const boost::property_tree::ptree& valhalla::config(const std::string& config_file_or_inline);
+};
+} // namespace
+
+namespace valhalla {
+const boost::property_tree::ptree& config(const std::string& config_file_or_inline = "") {
+  static config_singleton_t instance(config_file_or_inline);
+  return instance.config_;
+}
+} // namespace valhalla
+
+#endif
